requirements.txt
django==1.8.3
djangorestframework==3.2.2
# django-yearlessdate ???